alexa_controller/requirements.txt
boto3
ask_sdk_core
ask_sdk_model
requests
python-chess
Pillow
cairosvg